refactor: make util folder more local
ref #2318
