+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/selectors/util/mainProgrammingLanguageHelper.ts
