+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/selectors/util/suspiciousMetricsHelper.ts
@@ -1,5 +1,5 @@
-import { CodeMapNode, ColorRange } from "../../../../codeCharta.model"
-import { metricDescriptions } from "../../../../util/metric/metricDescriptions"
+import { CodeMapNode, ColorRange } from "../../../../../codeCharta.model"
+import { metricDescriptions } from "../../../../../util/metric/metricDescriptions"
 import { getAssociatedMetricThresholds } from "./getMetricThresholds"
 
 interface MetricValues {
